# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ragger",
    "description": "Local multi-workspace RAG server for codebases and docs, backed by Chroma and Ollama.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Check server and Ollama health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List indexed workspaces",
        "operationId": "list_workspaces_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace}/stats": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get workspace stats",
        "operationId": "workspace_stats_workspaces__workspace__stats_get",
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace}/status": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get workspace status",
        "operationId": "workspace_status_workspaces__workspace__status_get",
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/index": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Index or reindex a workspace",
        "operationId": "index_workspace_workspaces_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/search": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Search an indexed workspace",
        "operationId": "search_workspace_workspaces_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace}": {
      "delete": {
        "tags": [
          "workspaces"
        ],
        "summary": "Delete an indexed workspace",
        "operationId": "delete_workspace_workspaces__workspace__delete",
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DeleteWorkspaceResponse": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "workspace",
          "deleted"
        ],
        "title": "DeleteWorkspaceResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "persist_directory": {
            "type": "string",
            "title": "Persist Directory"
          },
          "workspace_count": {
            "type": "integer",
            "title": "Workspace Count"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "ollama": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ollama"
          }
        },
        "type": "object",
        "required": [
          "status",
          "persist_directory",
          "workspace_count",
          "embedding_model",
          "model",
          "ollama"
        ],
        "title": "HealthStatus"
      },
      "IndexRequest": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace",
            "default": "default"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "replace": {
            "type": "boolean",
            "title": "Replace",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "IndexRequest"
      },
      "IndexResponse": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace"
          },
          "root_path": {
            "type": "string",
            "title": "Root Path"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "indexed_extensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Indexed Extensions"
          },
          "last_indexed_at": {
            "type": "string",
            "title": "Last Indexed At"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "workspace",
          "root_path",
          "collection_name",
          "file_count",
          "chunk_count",
          "indexed_extensions",
          "last_indexed_at",
          "embedding_model",
          "model",
          "duration_seconds"
        ],
        "title": "IndexResponse"
      },
      "SearchHit": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "relative_path": {
            "type": "string",
            "title": "Relative Path"
          },
          "workspace": {
            "type": "string",
            "title": "Workspace"
          },
          "extension": {
            "type": "string",
            "title": "Extension"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "content_preview": {
            "type": "string",
            "title": "Content Preview"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "source",
          "relative_path",
          "workspace",
          "extension",
          "language",
          "content_preview",
          "content"
        ],
        "title": "SearchHit"
      },
      "SearchRequest": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace",
            "default": "default"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "workspace",
          "query",
          "results"
        ],
        "title": "SearchResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceStats": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace"
          },
          "root_path": {
            "type": "string",
            "title": "Root Path"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "indexed_extensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Indexed Extensions"
          },
          "last_indexed_at": {
            "type": "string",
            "title": "Last Indexed At"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "workspace",
          "root_path",
          "collection_name",
          "file_count",
          "chunk_count",
          "indexed_extensions",
          "last_indexed_at",
          "embedding_model",
          "model"
        ],
        "title": "WorkspaceStats"
      },
      "WorkspacesResponse": {
        "properties": {
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceStats"
            },
            "type": "array",
            "title": "Workspaces"
          }
        },
        "type": "object",
        "required": [
          "workspaces"
        ],
        "title": "WorkspacesResponse"
      }
    }
  }
}